requirements.txt
# Base ----------------------------------------
opencv-contrib-python
matplotlib
numpy
tornado
pynput
pillow